requirement.txt
python-dotenv
aiogram
openai==0.27.0
openai